+++ src/mqt/bench/targets/devices/iqm.py
@@ -10,7 +10,7 @@
 
 from __future__ import annotations
 
-from qiskit.circuit import Parameter
+from qiskit.circuit import IfElseOp, Parameter
 from qiskit.circuit.library import CZGate, Measure, RGate
 from qiskit.transpiler import InstructionProperties, Target
 
@@ -217,5 +217,8 @@ def _build_iqm_target(
     # === Add two-qubit gates ===
     cz_props = {(q1, q2): InstructionProperties(duration=twoq_duration, error=twoq_error) for q1, q2 in connectivity}
     target.add_instruction(CZGate(), cz_props)
+
+    # === Add control-flow ops ===
+    target.add_instruction(IfElseOp, name="if_else")
 
     return target