+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- ✨ Add support for compiling `if_test` for all gatesets and devices ([#805]) ([**@adnathanail**])
+
 ### Changed
 
 - 🏷️ Replace `mypy` with `ty` for static type checking ([#794]) ([**@burgholzer**])
@@ -93,6 +97,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 
 <!-- PR links -->
 
+[#805]: https://github.com/munich-quantum-toolkit/bench/pull/805
 [#794]: https://github.com/munich-quantum-toolkit/bench/pull/794
 [#731]: https://github.com/munich-quantum-toolkit/bench/pull/731
 [#709]: https://github.com/munich-quantum-toolkit/bench/pull/709
@@ -138,6 +143,7 @@ _📚 Refer to the [GitHub Release Notes] for previous changelogs._
 [**@CreativeBinBag**]: https://github.com/CreativeBinBag
 [**@denialhaag**]: https://github.com/denialhaag
 [**@soroushfathi**]: https://github.com/soroushfathi
+[**@adnathanail**]: https://github.com/adnathanail
 
 <!-- General links -->